pyproject.toml
[project]
name = "nbastats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-bigquery>=3.40.0",
    "google-cloud-storage>=3.8.0",
    "nba-api>=1.11.3",
]
